feat: simplify logger interface - let users provide their own
- Remove createLogger and createConsoleLogger exports from SDK
- Make logger optional with silent mode as default
- Remove console-logger.ts and logger.ts implementation files
- Update all tests to use mock logger implementations
- Remove LoggerConfig and LoggerFactory types
- Update documentation to show custom logger examples
- Bundle size unchanged: still ~165KB

BREAKING CHANGE: Users must now provide their own logger implementation
or omit the logger parameter for silent mode. The SDK no longer exports
createConsoleLogger or any logger factory functions.
