+++ tests/services/command-service.test.ts
@@ -1,12 +1,17 @@
 import { afterEach, beforeEach, describe, expect, it, type Mock, mock } from 'bun:test';
-import { createLogger } from '@/lib/logger';
 import * as commandService from '@/services/command-service';
 import type { Logger } from '@/types/logger';
 
 describe('command-service', () => {
   describe('executeCommand', () => {
     let mockFetch: Mock<typeof fetch>;
-    const mockLogger: Logger = createLogger();
+    const mockLogger: Logger = {
+      info: () => {},
+      error: () => {},
+      debug: () => {},
+      warn: () => {},
+      level: 'silent',
+    };
     const testAppName = 'test-app';
 
     beforeEach(() => {
